requirements.txt
groq
python-dotenv
textblob
-e .